requirements.txt
streamlit-1.27.0-py2.py3-none-any.whl